requirements.txt
librosa
pygame
PyQt6
PyOpenGL
numpy
opencv-python
python-osc
SpoutGL
sacn
dearpygui
pedalboard